+++ CHANGELOG.md
@@ -1,9 +1,19 @@
 # Change Log
 
-All notable changes to the "code-tutor" extension will be documented in this file.
-
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
-
-## [Unreleased]
+## [1.0.0]
 
 - Initial release
+
+## [1.0.1]
+
+- Update Language Models
+
+## [1.0.2]
+
+- Add C and C++ functionality
+
+## [1.0.3]
+
+- Added LICENCE file
+- changed code structure
+- updated extension-quickstart file